[docs] Split test ad networks (#12878)
* [docs] Split test ad networks

* Fix feedback

* Fix prettier

* small edits
